requirements.txt
# Core Django
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==2023.10
django-filter==23.3
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.0.0

# Environment variables
python-decouple==3.8

# Static files
whitenoise==6.5.0

# Security
bcrypt==4.0.1
django-axes==6.0.0

# API
django-cors-headers==4.2.0

# Utilities
Pillow==10.0.0
python-dateutil==2.8.2
pandas==2.1.1
openpyxl==3.1.2

# Development
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0